doDelete, doModify 권한 체크 기능 추가

+++ src/main/java/com/gbr/exam/sbam/controller/UserArticleController.java
@@ -64,11 +64,28 @@ public class UserArticleController {
 
 	@RequestMapping("/user/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "!! 로그인 후 이용 가능합니다 !!");
+		}
+		
 		Article article = userArticleService.getArticle(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다 :(", id), id);
+		} 
+		// 삭제 권한 체크
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 게시물에 대한 삭제 권한이 없습니다.", id));
 		}
 
 		userArticleService.deleteArticle(id);
@@ -78,16 +95,34 @@ public class UserArticleController {
 	
 	@RequestMapping("/user/article/doModify")
 	@ResponseBody
-	public ResultData<Integer> doModify(int id, String title, String body) {
+	public ResultData doModify(HttpSession httpSession, int id, String title, String body) {
+
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
+		
 		Article article = userArticleService.getArticle(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다 :(", id), id);
 		}
 
-		userArticleService.modifyArticle(id, title, body);
+		ResultData actorCanModifyRd = userArticleService.actorCanModify(loginedMemberId, article);
 
-		return ResultData.from("S-1", Ut.f("%d번 게시물이 수정되었습니다. :)", id), id);
+		if (actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
+
+		return userArticleService.modifyArticle(id, title, body);
+
 	}
 
 	@RequestMapping("/user/article/getArticle")
